# server/app/api/swagger.yml
openapi: "3.0.3"
info:
  description: "This is the Databus API documentation."
  version: "2.1.0"
  title: "Databus API"
components:
  securitySchemes:
    ApiKeyAuth:        
      type: apiKey
      in: header       
      name: X-API-KEY  
paths:
...
  /api/account/webid/add:
    post:
      summary: "Add WebId"
      description: "Links an external WebId to the authenticated Databus Account."
      operationId: "account-webid-link"
      security:
        - ApiKeyAuth: []
      tags:
        - Account
      parameters:
        - name: uri
          in: query
          description: the WebId uri to link
          schema:
            type: string
          required: true
          example: 'https://holycrab13.github.io/webid.ttl'
      responses:
        "200":
          description: "OK"
        "403":
          description: "Error: Forbidden"
          content:
            text/plain:
              schema:
                type: string
                example: "Unable to find valid backlink in WebId document."
  /api/account/webid/remove:
    post:
      summary: "Remove WebId"
      description: "Removes an external WebId from the authenticated Databus Account."
      operationId: "account-webid-remove"
      security:
        - ApiKeyAuth: []
      tags:
        - Account
      parameters:
        - name: uri
          in: query
          description: the WebId uri to remove
          schema:
            type: string
          required: true
          example: 'https://holycrab13.github.io/webid.ttl'
      responses:
        "200":
          description: "OK"
          content:
            text/plain:
              schema:
                type: string
                example: "WebId removed from account."
  /api/account/api-key/create:
    post:
      summary: "Create API key"
      description: "Create a new API key."
      operationId: "account-apikey-create"
      security:
        - ApiKeyAuth: []
      tags:
        - Account
      parameters:
        - name: name
          in: query
          description: the desired name of the api key
          schema:
            type: string
          required: true
          example: 'default'
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: object
                example: '{
                    "name": "default",
                    "key": "2df02a7f-32c4-46cd-9c6e-bd80c72e6e48"
                  }'
        "400":
          description: "Error. Bad Request"
          content:
            text/plain:
              schema:
                type: string
                example: "Failed to create API key. You might already have an API key with that name."
        "403":
          description: "Error. Forbidden"
          content:
            text/plain:
              schema:
                type: string
                example: "Account name is missing. Please claim an account name first."
  /api/account/api-key/delete:
    post:
      summary: "Delete API key"
      description: "Delete an existing API key."
      operationId: "account-apikey-delete"
      security:
        - ApiKeyAuth: []
      tags:
        - Account
      parameters:
        - name: name
          in: query
          description: the name of the api key to delete
          schema:
            type: string
          required: true
          example: 'default'
      responses:
        "200":
          description: "OK"
        "204":
          description: "Api-Key didn't exist"
...